+++ glean/src/core/glean.ts
@@ -509,14 +509,19 @@ if (typeof window !== "undefined" && typeof window.sessionStorage !== "undefined
     setLogPings: (flag: boolean) => {
       setDebugOptionInSessionStorage(DebugOption.LogPings, flag);
       Glean.setLogPings(flag);
+      console.log("Pings will be logged to the console until this tab is closed.");
     },
     setDebugViewTag: (value: string) => {
       setDebugOptionInSessionStorage(DebugOption.DebugTag, value);
       Glean.setDebugViewTag(value);
+      console.log(
+        "Pings will be sent to the Debug Ping Viewer until this tab is closed. Pings can be found here: https://debug-ping-preview.firebaseapp.com/."
+      );
     },
     setSourceTags: (value: string[]) => {
       setDebugOptionInSessionStorage(DebugOption.SourceTags, value);
       Glean.setSourceTags(value);
+      console.log("Pings will be given the specified tags until the tab is closed.");
     }
   };
 }