+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 [#1835](https://github.com/mozilla/glean.js/pull/1835): Added support for automatic page load instrumentation.
 
+* [#1846](https://github.com/mozilla/glean.js/pull/1846): Add logging messages when using the debugging APIs from the browser console.
+
 # v4.0.0-pre.1 (2023-12-01)
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v4.0.0-pre.0...v4.0.0-pre.1)